requirements.txt
ollama
pydantic
pytest

psycopg

python-dotenv

# GUI Stuff -----
# PySide6